moved changes from save-osv
