+++ save-cosv/src/main/kotlin/com/saveourtool/save/backend/service/IVulnerabilityService.kt
@@ -3,6 +3,7 @@
 package com.saveourtool.save.backend.service
 
 import com.saveourtool.save.entities.vulnerabilities.Vulnerability
+import com.saveourtool.save.entities.vulnerability.ProposeSaveOsvRequest
 import com.saveourtool.save.entities.vulnerability.VulnerabilityDto
 import org.springframework.security.core.Authentication
 
@@ -25,5 +26,15 @@ interface IVulnerabilityService {
     fun save(
         vulnerabilityDto: VulnerabilityDto,
         authentication: Authentication,
+    ): Vulnerability
+
+    /**
+     * @param proposeSaveOsvRequest a request to create a new vulnerability meta
+     * @param creatorUserId id of user who proposes OSV
+     * @return saved [Vulnerability]
+     */
+    fun save(
+        proposeSaveOsvRequest: ProposeSaveOsvRequest,
+        creatorUserId: Long,
     ): Vulnerability
 }
