+++ save-cosv/src/main/kotlin/com/saveourtool/save/cosv/service/CosvService.kt
@@ -1,5 +1,6 @@
 package com.saveourtool.save.cosv.service
 
+import com.saveourtool.osv4k.OsvSchema
 import com.saveourtool.save.backend.service.IVulnerabilityService
 import com.saveourtool.save.cosv.processor.CosvProcessorHolder
 import com.saveourtool.save.cosv.repository.CosvRepository
@@ -8,6 +9,9 @@ import com.saveourtool.save.entities.vulnerability.*
 import com.saveourtool.save.utils.*
 
 import com.saveourtool.osv4k.RawOsvSchema
+import com.saveourtool.osv4k.Reference
+import com.saveourtool.osv4k.ReferenceType
+import com.saveourtool.save.info.UserInfo
 import org.springframework.security.core.Authentication
 import org.springframework.stereotype.Service
 import reactor.core.publisher.Flux
@@ -16,6 +20,8 @@ import reactor.kotlin.core.publisher.toMono
 
 import java.io.InputStream
 
+import kotlinx.datetime.LocalDateTime
+import kotlinx.datetime.toKotlinLocalDateTime
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.*
 import kotlinx.serialization.serializer
@@ -110,4 +116,53 @@ class CosvService(
         .flatMap { vulnerability ->
             cosvRepository.findLatestById(vulnerability.name, serializer<RawOsvSchema>())
         }
+
+    fun createNew(
+        proposeSaveOsvRequest: ProposeSaveOsvRequest,
+        creatorUserId: Long,
+    ): Mono<String> = blockingToMono { vulnerabilityService.save(proposeSaveOsvRequest, creatorUserId) }
+        .flatMap { vulnerabilityMeta ->
+            val osv = OsvSchema<Unit, Unit, Unit, Unit>(
+                schemaVersion = "1.5.0",
+                id = vulnerabilityMeta.name,
+                modified = vulnerabilityMeta.createDate.orNotFound {
+                    "CreationDate is not provided on vulnerability meta ${vulnerabilityMeta.name}"
+                }.toKotlinLocalDateTime(),
+                summary = proposeSaveOsvRequest.shortDescription,
+                details = proposeSaveOsvRequest.description,
+                aliases = proposeSaveOsvRequest.vulnerabilityIdentifier?.let { listOf(it) },
+                references = proposeSaveOsvRequest.relatedLink?.let { listOf(Reference(ReferenceType.WEB, it)) }
+            )
+            cosvRepository.save(osv, serializer<OsvSchema<Unit, Unit, Unit, Unit>>())
+                .map { osv.id }
+        }
+
+    companion object {
+        private fun RawOsvSchema.toSaveVulnerabilityDto(): VulnerabilityDto = VulnerabilityDto(
+            name = "TO_GENERATE",  // will be generated on saving to database
+            vulnerabilityIdentifier = id,
+            progress = 0,
+            projects = emptyList(),
+            description = details,
+            shortDescription = summary.orEmpty(),
+            relatedLink = null,
+            language = VulnerabilityLanguage.OTHER,
+            userInfo = UserInfo(name = ""),  // will be set on saving to database
+            organization = null,
+            dates = buildList {
+                add(modified.asVulnerabilityDateDto(VulnerabilityDateType.CVE_UPDATED))
+                published?.asVulnerabilityDateDto(VulnerabilityDateType.INTRODUCED)?.run { add(this) }
+                withdrawn?.asVulnerabilityDateDto(VulnerabilityDateType.FIXED)?.run { add(this) }
+            },
+            participants = emptyList(),
+            status = VulnerabilityStatus.CREATED,
+            tags = setOf("osv-schema")
+        )
+
+        private fun LocalDateTime.asVulnerabilityDateDto(type: VulnerabilityDateType) = VulnerabilityDateDto(
+            date = this,
+            type = type,
+            vulnerabilityName = "NOT_USED_ON_SAVE",
+        )
+    }
 }
