--- save-osv/src/main/kotlin/com/saveourtool/save/backend/service/IVulnerabilityService.kt
@@ -1,40 +1,0 @@
-@file:Suppress("FILE_NAME_INCORRECT")
-
-package com.saveourtool.save.backend.service
-
-import com.saveourtool.save.entities.vulnerabilities.Vulnerability
-import com.saveourtool.save.entities.vulnerability.ProposeSaveOsvRequest
-import com.saveourtool.save.entities.vulnerability.VulnerabilityDto
-import org.springframework.security.core.Authentication
-
-/**
- * Interface for VulnerabilityService
- */
-@Suppress("CLASS_NAME_INCORRECT")
-interface IVulnerabilityService {
-    /**
-     * @param name name of vulnerability
-     * @return vulnerability by name
-     */
-    fun findByName(name: String): Vulnerability?
-
-    /**
-     * @param vulnerabilityDto dto of new vulnerability
-     * @param authentication auth info of a current user
-     * @return saved [Vulnerability]
-     */
-    fun save(
-        vulnerabilityDto: VulnerabilityDto,
-        authentication: Authentication,
-    ): Vulnerability
-
-    /**
-     * @param proposeSaveOsvRequest a request to create a new vulnerability meta
-     * @param creatorUserId id of user who proposes OSV
-     * @return saved [Vulnerability]
-     */
-    fun save(
-        proposeSaveOsvRequest: ProposeSaveOsvRequest,
-        creatorUserId: Long,
-    ): Vulnerability
-}
