--- save-osv/src/main/kotlin/com/saveourtool/save/osv/OsvConfiguration.kt
@@ -1,11 +1,0 @@
-package com.saveourtool.save.osv
-
-import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
-
-/**
- * Configuration for Vulnerability
- */
-@Configuration
-@ComponentScan
-class OsvConfiguration
