--- save-osv/src/main/kotlin/com/saveourtool/save/osv/storage/OsvKey.kt
@@ -1,12 +1,0 @@
-package com.saveourtool.save.osv.storage
-
-import kotlinx.datetime.LocalDateTime
-
-/**
- * @property id
- * @property modified
- */
-data class OsvKey(
-    val id: String,
-    val modified: LocalDateTime,
-)
